Revert "support uploading multiple pictures at once"
This reverts commit eb3d9e7cde3ac37a9ae11e661592eb50eb74f1cb.

I had made this change for a user who requested this ability.  But, when users upload a large number of images at once it creates a large payload which tends to exceed the Nginx max-request-size.  I think it is sane to keep the max request size smaller (currently 25mb), and Haven is intended for status-update type use, instead of photo-album type use.  So I'm reverting this to prevent unexpected "request size too big" errors.

Better support for large numbers of images may be incorporated into Haven later.